Consolidate storage validation into conformance tests
