--- tests/Immediate.Jobs.StorageTests/RelationalStorageMatrixTests.cs
@@ -1,1042 +1,0 @@
-using System.Globalization;
-using System.Text.RegularExpressions;
-using Immediate.Jobs.EntityFrameworkCore;
-using Immediate.Jobs.LinqToDB;
-using Immediate.Jobs.Shared.Apis;
-using Immediate.Jobs.Shared.Internals;
-using Immediate.Jobs.Shared.Storage;
-using LinqToDB;
-using LinqToDB.Data;
-using Microsoft.Data.Sqlite;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.Extensions.Time.Testing;
-
-namespace Immediate.Jobs.StorageTests;
-
-[Collection(StorageContainerFixtureGroup.Name)]
-public sealed class RelationalStorageMatrixTests(StorageContainers containers)
-{
-	public static TheoryData<DatabaseKind, AdapterKind> Matrix => new()
-	{
-		{ DatabaseKind.Sqlite, AdapterKind.EntityFrameworkCore },
-		{ DatabaseKind.Sqlite, AdapterKind.LinqToDB },
-		{ DatabaseKind.PostgreSql, AdapterKind.EntityFrameworkCore },
-		{ DatabaseKind.PostgreSql, AdapterKind.LinqToDB },
-		{ DatabaseKind.SqlServer, AdapterKind.EntityFrameworkCore },
-		{ DatabaseKind.SqlServer, AdapterKind.LinqToDB },
-	};
-
-	public static TheoryData<AdapterKind> SqlServerAdapters =>
-	[
-		AdapterKind.EntityFrameworkCore,
-		AdapterKind.LinqToDB,
-	];
-
-	public static TheoryData<DatabaseKind> LinqToDbDatabases =>
-	[
-		DatabaseKind.Sqlite,
-		DatabaseKind.PostgreSql,
-		DatabaseKind.SqlServer,
-	];
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task AdapterHandlesContentionAndLeaseRecovery(DatabaseKind database, AdapterKind adapter)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		var first = fixture.CreateStorage();
-		var second = fixture.CreateStorage();
-		var now = fixture.TimeProvider.GetUtcNow();
-		foreach (var index in Enumerable.Range(0, 24))
-			await first.EnqueueAsync(CreateJob(string.Create(CultureInfo.InvariantCulture, $"contended-{index}"), now), cancellationToken);
-
-		var claims = await Task.WhenAll(
-			first.AcquireDueJobsAsync(CreateRequest("node-a", 24), cancellationToken).AsTask(),
-			second.AcquireDueJobsAsync(CreateRequest("node-b", 24), cancellationToken).AsTask()
-		);
-		var claimed = claims.SelectMany(static claim => claim).ToArray();
-		Assert.Equal(24, claimed.Select(job => job.Id).Distinct().Count());
-		foreach (var job in claimed)
-			await first.CompleteAsync(job.Id, job.Attempt, job.WorkerId!, cancellationToken);
-
-		var leased = CreateJob("leased", now.AddMinutes(1));
-		await first.EnqueueAsync(leased, cancellationToken);
-		fixture.TimeProvider.Advance(TimeSpan.FromMinutes(1));
-		var original = Assert.Single(await first.AcquireDueJobsAsync(CreateRequest("node-a", 1), cancellationToken));
-		await first.SetExecutionTelemetryAsync(
-			leased.Id,
-			1,
-			"node-a",
-			"4bf92f3577b34da6a3ce929d0e0e4736",
-			"00f067aa0ba902b7",
-			fixture.TimeProvider.GetUtcNow(),
-			cancellationToken
-		);
-		fixture.TimeProvider.Advance(TimeSpan.FromMinutes(2));
-		var recovered = Assert.Single(await second.AcquireDueJobsAsync(CreateRequest("node-a", 1), cancellationToken));
-		Assert.Equal("leased", recovered.Id);
-		Assert.Equal(2, recovered.Attempt);
-		Assert.Equal("node-a", recovered.WorkerId);
-		Assert.Null(recovered.ExecutionTraceId);
-		Assert.Null(recovered.ExecutionSpanId);
-		Assert.Null(recovered.ExecutionStartedAt);
-
-		_ = await Assert.ThrowsAsync<ImmediateJobException>(
-			() => first.RenewLeaseAsync(leased.Id, original.Attempt, "node-a", TimeSpan.FromMinutes(1), cancellationToken).AsTask()
-		);
-		_ = await Assert.ThrowsAsync<ImmediateJobException>(() => first.SetExecutionTelemetryAsync(
-			leased.Id,
-			original.Attempt,
-			"node-a",
-			"stale",
-			"stale",
-			fixture.TimeProvider.GetUtcNow(),
-			cancellationToken
-		).AsTask());
-		_ = await Assert.ThrowsAsync<ImmediateJobException>(
-			() => first.CompleteAsync(leased.Id, original.Attempt, "node-a", cancellationToken).AsTask()
-		);
-		_ = await Assert.ThrowsAsync<ImmediateJobException>(() => first.FailAsync(
-			leased.Id,
-			original.Attempt,
-			"node-a",
-			"stale",
-			nextRetryAt: null,
-			cancellationToken
-		).AsTask());
-
-		var executions = await first.QueryJobExecutionsAsync(new() { JobId = leased.Id }, cancellationToken);
-		Assert.Collection(
-			executions,
-			execution =>
-			{
-				Assert.Equal(2, execution.Attempt);
-				Assert.Equal(JobExecutionState.Active, execution.State);
-			},
-			execution =>
-			{
-				Assert.Equal(1, execution.Attempt);
-				Assert.Equal(JobExecutionState.Interrupted, execution.State);
-				Assert.Equal(original.LeaseExpiresAt, execution.CompletedAt);
-				Assert.Equal("4bf92f3577b34da6a3ce929d0e0e4736", execution.ExecutionTraceId);
-			}
-		);
-	}
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task ReclaimedLeaseRejectsStaleCompletionAndItsBufferedContinuation(
-		DatabaseKind database,
-		AdapterKind adapter
-	)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		var first = fixture.CreateStorage();
-		var second = fixture.CreateStorage();
-		var firstGraph = Assert.IsAssignableFrom<IJobGraphStorage>(first);
-		var now = fixture.TimeProvider.GetUtcNow();
-		await first.EnqueueAsync(CreateJob("reclaimed-parent", now), cancellationToken);
-		_ = Assert.Single(await first.AcquireDueJobsAsync(CreateRequest("node-a", 1), cancellationToken));
-		fixture.TimeProvider.Advance(TimeSpan.FromMinutes(2));
-		var reclaimed = Assert.Single(await second.AcquireDueJobsAsync(CreateRequest("node-b", 1), cancellationToken));
-
-		var exception = await Assert.ThrowsAsync<ImmediateJobException>(() =>
-			firstGraph.CompleteWithContinuationsAsync(
-				"reclaimed-parent",
-				1, "node-a",
-				[new()
-				{
-					Job = CreateJob("stale-continuation", fixture.TimeProvider.GetUtcNow()),
-					Options = ContinuationOptions.Detached,
-				}],
-				cancellationToken
-			).AsTask()
-		);
-
-		Assert.Contains("does not own active job", exception.Message, StringComparison.Ordinal);
-		Assert.Empty(await first.QueryJobsAsync(new() { Id = "stale-continuation" }, cancellationToken));
-		Assert.Equal("node-b", reclaimed.WorkerId);
-		Assert.Equal(JobState.Active, (await first.GetJobStatusAsync("reclaimed-parent", cancellationToken))!.State);
-	}
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task ConcurrentLegacyRetriesTreatSyntheticExecutionInsertAsARace(
-		DatabaseKind database,
-		AdapterKind adapter
-	)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		var first = fixture.CreateStorage();
-		var second = fixture.CreateStorage();
-		var now = fixture.TimeProvider.GetUtcNow();
-		await first.EnqueueAsync(CreateJob("legacy-retry", now) with
-		{
-			State = JobState.Failed,
-			Attempt = 1,
-			CompletedAt = now,
-			LastError = "legacy failure",
-		}, cancellationToken);
-		var start = new TaskCompletionSource(TaskCreationOptions.RunContinuationsAsynchronously);
-
-		async Task<Exception?> RetryAsync(IJobStorage storage)
-		{
-			await start.Task;
-			return await Record.ExceptionAsync(() => storage.RetryAsync("legacy-retry", cancellationToken).AsTask());
-		}
-
-		var retries = new[] { RetryAsync(first), RetryAsync(second) };
-		start.SetResult();
-		var exceptions = await Task.WhenAll(retries);
-
-		_ = Assert.Single(exceptions, static exception => exception is null);
-		_ = Assert.IsType<ImmediateJobException>(Assert.Single(exceptions.OfType<Exception>()));
-		var execution = Assert.Single(await first.QueryJobExecutionsAsync(
-			new() { JobId = "legacy-retry" },
-			cancellationToken
-		));
-		Assert.True(execution.IsSynthetic);
-	}
-
-	[Theory]
-	[MemberData(nameof(LinqToDbDatabases))]
-	public async Task LinqCompletionSurvivesSustainedBatchHeaderContention(DatabaseKind database)
-	{
-		const int JobCount = 24;
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(
-			database,
-			AdapterKind.LinqToDB,
-			cancellationToken
-		);
-		var storage = fixture.GraphStorage;
-		var replica = Assert.IsAssignableFrom<IJobStorageReplica>(storage);
-		var now = fixture.TimeProvider.GetUtcNow();
-		var jobs = Enumerable.Range(0, JobCount)
-			.Select(index =>
-			{
-				var id = string.Create(CultureInfo.InvariantCulture, $"contended-completion-{index}");
-				return CreateJob(id, now) with { BatchId = "contended-batch" };
-			})
-			.ToArray();
-		await storage.EnqueueBatchAsync(new()
-		{
-			Id = "contended-batch",
-			CreatedAt = now,
-			TotalJobs = JobCount,
-			PendingCount = JobCount,
-			State = BatchState.Executing,
-		}, jobs, [], cancellationToken);
-		Assert.Equal(JobCount, (await replica.AcquireJobsAsync(
-			[.. jobs.Select(static job => job.Id)],
-			"worker",
-			TimeSpan.FromMinutes(1),
-			cancellationToken
-		)).Count);
-		var start = new TaskCompletionSource(TaskCreationOptions.RunContinuationsAsynchronously);
-		var completions = jobs.Select(async job =>
-		{
-			await start.Task;
-			await storage.CompleteAsync(job.Id, 1, "worker", cancellationToken);
-		}).ToArray();
-
-		start.SetResult();
-		await Task.WhenAll(completions);
-
-		var batch = Assert.IsType<BatchStatus>(await storage.GetBatchStatusAsync("contended-batch", cancellationToken));
-		Assert.Equal(BatchState.Succeeded, batch.State);
-		Assert.Equal(JobCount, batch.Succeeded);
-	}
-
-	[Theory]
-	[MemberData(nameof(LinqToDbDatabases))]
-	public async Task LinqRecurringUpsertSurvivesConcurrentMaterialization(DatabaseKind database)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(
-			database,
-			AdapterKind.LinqToDB,
-			cancellationToken
-		);
-		var storage = fixture.RecurringStorage;
-		var now = fixture.TimeProvider.GetUtcNow();
-		var schedule = new RecurringJobSchedule
-		{
-			Name = "racing-upsert",
-			JobName = "matrix-test",
-			Cron = "0 * * * *",
-			TimeZone = "UTC",
-			IsCodeDefined = false,
-			NextRunAt = now,
-		};
-		await storage.UpsertRecurringAsync(schedule, cancellationToken);
-
-		for (var iteration = 0; iteration < 20; iteration++)
-		{
-			var current = Assert.Single((await storage.GetMonitoringSnapshotAsync(cancellationToken)).Recurring);
-			var requested = schedule with
-			{
-				Cron = iteration % 2 == 0 ? "15 * * * *" : "45 * * * *",
-				NextRunAt = now.AddDays(iteration + 1),
-			};
-			var occurrence = CreateJob(
-				string.Create(CultureInfo.InvariantCulture, $"racing-occurrence-{iteration}"),
-				now
-			) with
-			{
-				RecurringKey = string.Create(
-					CultureInfo.InvariantCulture,
-					$"racing-upsert:{current.NextRunAt.UtcTicks}"
-				),
-			};
-			var start = new TaskCompletionSource(TaskCreationOptions.RunContinuationsAsynchronously);
-			async Task MaterializeAsync()
-			{
-				await start.Task;
-				_ = await storage.MaterializeRecurringAsync(
-					current,
-					occurrence,
-					current.NextRunAt.AddMinutes(1),
-					cancellationToken
-				);
-			}
-
-			async Task UpsertAsync()
-			{
-				await start.Task;
-				await storage.UpsertRecurringAsync(requested, cancellationToken);
-			}
-
-			var materialization = MaterializeAsync();
-			var upsert = UpsertAsync();
-
-			start.SetResult();
-			await Task.WhenAll(materialization, upsert);
-
-			var persisted = Assert.Single((await storage.GetMonitoringSnapshotAsync(cancellationToken)).Recurring);
-			Assert.Equal(requested.Cron, persisted.Cron);
-			Assert.Equal(requested.NextRunAt, persisted.NextRunAt);
-		}
-	}
-
-	[Theory]
-	[MemberData(nameof(SqlServerAdapters))]
-	public async Task SqlServerCollationIsUsedConsistentlyForGroupState(AdapterKind adapter)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(
-			DatabaseKind.SqlServer,
-			adapter,
-			cancellationToken
-		);
-		var storage = fixture.Storage;
-		var now = fixture.TimeProvider.GetUtcNow();
-		await storage.EnqueueAsync(
-			CreateJob("tenant-first", now) with { GroupId = "Tenant" },
-			cancellationToken
-		);
-		await storage.EnqueueAsync(
-			CreateJob("tenant-second", now) with
-			{
-				GroupId = "tenant",
-				CreatedAt = now.AddTicks(1),
-			},
-			cancellationToken
-		);
-		var request = CreateRequest("fair-worker", 1) with
-		{
-			FairQueues = new FairQueuePolicy { ConcurrencyShareThreshold = 0.10, MinInflightForNoisy = 30, GroupRoundRobin = true },
-		};
-		var first = Assert.Single(await storage.AcquireDueJobsAsync(request, cancellationToken));
-		Assert.Equal("tenant-first", first.Id);
-		await storage.CompleteAsync(first.Id, 1, "fair-worker", cancellationToken);
-		await storage.EnqueueAsync(
-			CreateJob("quiet-first", now) with
-			{
-				GroupId = "quiet",
-				CreatedAt = now.AddTicks(2),
-			},
-			cancellationToken
-		);
-
-		var second = Assert.Single(await storage.AcquireDueJobsAsync(request, cancellationToken));
-
-		Assert.Equal("quiet-first", second.Id);
-	}
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task RetriedParentDoesNotSettleContinuationEdgeTwice(
-		DatabaseKind database,
-		AdapterKind adapter
-	)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		var storage = fixture.GraphStorage;
-		var replica = Assert.IsAssignableFrom<IJobStorageReplica>(storage);
-		var now = fixture.TimeProvider.GetUtcNow();
-		var retriedParent = CreateJob("retried-parent", now);
-		var otherParent = CreateJob("other-parent", now.AddTicks(1)) with { DueAt = now };
-		var child = CreateJob("retry-child", now.AddTicks(2)) with { DueAt = now };
-		await storage.EnqueueAsync(retriedParent, cancellationToken);
-		await storage.EnqueueAsync(otherParent, cancellationToken);
-		await storage.EnqueueContinuationAsync(child,
-		[
-			new()
-			{
-				ChildJobId = child.Id,
-				ParentJobId = retriedParent.Id,
-				Trigger = ContinuationTrigger.Complete,
-			},
-			new()
-			{
-				ChildJobId = child.Id,
-				ParentJobId = otherParent.Id,
-				Trigger = ContinuationTrigger.Complete,
-			},
-		], cancellationToken);
-		_ = Assert.Single(await replica.AcquireJobsAsync(
-			[retriedParent.Id],
-			"worker",
-			TimeSpan.FromMinutes(1),
-			cancellationToken
-		));
-		await storage.FailAsync(retriedParent.Id, 1, "worker", "expected", nextRetryAt: null, cancellationToken);
-		await storage.RetryAsync(retriedParent.Id, cancellationToken);
-		_ = Assert.Single(await replica.AcquireJobsAsync(
-			[retriedParent.Id],
-			"worker",
-			TimeSpan.FromMinutes(1),
-			cancellationToken
-		));
-
-		await storage.CompleteAsync(retriedParent.Id, 2, "worker", cancellationToken);
-
-		Assert.Equal(
-			JobState.AwaitingContinuation,
-			(await storage.GetJobStatusAsync(child.Id, cancellationToken))!.State
-		);
-		_ = Assert.Single(await replica.AcquireJobsAsync(
-			[otherParent.Id],
-			"worker",
-			TimeSpan.FromMinutes(1),
-			cancellationToken
-		));
-		await storage.CompleteAsync(otherParent.Id, 1, "worker", cancellationToken);
-		Assert.Equal(JobState.Pending, (await storage.GetJobStatusAsync(child.Id, cancellationToken))!.State);
-	}
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task ConcurrentCompletionAndContinuationInsertCannotStrandChild(
-		DatabaseKind database,
-		AdapterKind adapter
-	)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		var storage = fixture.GraphStorage;
-		var replica = Assert.IsAssignableFrom<IJobStorageReplica>(storage);
-		var now = fixture.TimeProvider.GetUtcNow();
-		for (var index = 0; index < 20; index++)
-		{
-			var parentId = string.Create(CultureInfo.InvariantCulture, $"racing-parent-{index}");
-			var childId = string.Create(CultureInfo.InvariantCulture, $"racing-child-{index}");
-			var parent = CreateJob(parentId, now.AddTicks(index * 2)) with { DueAt = now };
-			var child = CreateJob(childId, now.AddTicks((index * 2) + 1)) with { DueAt = now };
-			await storage.EnqueueAsync(parent, cancellationToken);
-			_ = Assert.Single(await replica.AcquireJobsAsync(
-				[parent.Id],
-				"worker",
-				TimeSpan.FromMinutes(1),
-				cancellationToken
-			));
-			var start = new TaskCompletionSource(TaskCreationOptions.RunContinuationsAsynchronously);
-			var completion = Task.Run(async () =>
-			{
-				await start.Task;
-				await storage.CompleteAsync(parent.Id, 1, "worker", cancellationToken);
-			}, cancellationToken);
-			var insertion = Task.Run(async () =>
-			{
-				await start.Task;
-				await storage.EnqueueContinuationAsync(child, [new()
-				{
-					ChildJobId = child.Id,
-					ParentJobId = parent.Id,
-					Trigger = ContinuationTrigger.Complete,
-				}], cancellationToken);
-			}, cancellationToken);
-
-			start.SetResult();
-			await Task.WhenAll(completion, insertion);
-
-			Assert.Equal(JobState.Pending, (await storage.GetJobStatusAsync(child.Id, cancellationToken))!.State);
-		}
-	}
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task DynamicAdditionsRejectInvalidBatchIdsBeforeMutation(
-		DatabaseKind database,
-		AdapterKind adapter
-	)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		var storage = fixture.GraphStorage;
-		var replica = Assert.IsType<IJobStorageReplica>(storage, exactMatch: false);
-		var now = fixture.TimeProvider.GetUtcNow();
-		var current = CreateJob("dynamic-current", now) with { BatchId = "dynamic-batch" };
-		await storage.EnqueueBatchAsync(new()
-		{
-			Id = "dynamic-batch",
-			CreatedAt = now,
-			TotalJobs = 1,
-			PendingCount = 1,
-			State = BatchState.Executing,
-		}, [current], [], cancellationToken);
-		_ = Assert.Single(await replica.AcquireJobsAsync(
-			[current.Id],
-			"worker",
-			TimeSpan.FromMinutes(1),
-			cancellationToken
-		));
-
-		_ = await Assert.ThrowsAsync<ImmediateJobException>(() => storage.CompleteWithContinuationsAsync(
-			current.Id,
-			1, "worker",
-			[new()
-			{
-				Job = CreateJob("detached-with-batch", now.AddTicks(1)) with { BatchId = "dynamic-batch" },
-				Options = ContinuationOptions.Detached,
-			}],
-			cancellationToken
-		).AsTask());
-		_ = await Assert.ThrowsAsync<ImmediateJobException>(() => storage.CompleteWithContinuationsAsync(
-			current.Id,
-			1, "worker",
-			[new()
-			{
-				Job = CreateJob("tracked-with-wrong-batch", now.AddTicks(2)) with { BatchId = "other-batch" },
-				Options = ContinuationOptions.BesideContinuations,
-			}],
-			cancellationToken
-		).AsTask());
-		_ = await Assert.ThrowsAsync<ImmediateJobException>(() => storage.AddBatchJobAsync(
-			current.Id,
-			1,
-			CreateJob("added-with-wrong-batch", now.AddTicks(3)) with { BatchId = "other-batch" },
-			ContinuationOptions.BesideContinuations,
-			cancellationToken
-		).AsTask());
-
-		var batch = Assert.IsType<BatchStatus>(await storage.GetBatchStatusAsync("dynamic-batch", cancellationToken));
-		Assert.Equal(1, batch.Total);
-		Assert.Equal(1, batch.Remaining);
-		Assert.Equal(JobState.Active, (await storage.GetJobStatusAsync(current.Id, cancellationToken))!.State);
-		foreach (var id in new[] { "detached-with-batch", "tracked-with-wrong-batch", "added-with-wrong-batch" })
-			Assert.Empty(await storage.QueryJobsAsync(new() { Id = id }, cancellationToken));
-	}
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task DeletingStandaloneContinuationRemovesIncomingEdges(
-		DatabaseKind database,
-		AdapterKind adapter
-	)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		var storage = fixture.GraphStorage;
-		var now = fixture.TimeProvider.GetUtcNow();
-		var parent = CreateJob("delete-parent", now) with { State = JobState.Succeeded, CompletedAt = now };
-		var child = CreateJob("delete-child", now) with { State = JobState.Succeeded, CompletedAt = now };
-		await storage.EnqueueAsync(parent, cancellationToken);
-		await storage.EnqueueContinuationAsync(child, [new()
-		{
-			ChildJobId = child.Id,
-			ParentJobId = parent.Id,
-			Trigger = ContinuationTrigger.Success,
-		}], cancellationToken);
-
-		_ = Assert.Single(await ((IJobGraphStorageReplica)storage).GetIncomingEdgesAsync([child.Id], cancellationToken));
-		await storage.DeleteAsync(child.Id, cancellationToken);
-
-		Assert.Null(await storage.GetJobStatusAsync(child.Id, cancellationToken));
-		Assert.Empty(await ((IJobGraphStorageReplica)storage).GetIncomingEdgesAsync([child.Id], cancellationToken));
-	}
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task CancelBatchMemberPropagatesContinuations(DatabaseKind database, AdapterKind adapter)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		var storage = fixture.GraphStorage;
-		var now = fixture.TimeProvider.GetUtcNow();
-		var parent = CreateJob("cancel-parent", now) with { BatchId = "cancel-member-batch" };
-		var successChild = CreateJob("cancel-success-child", now) with
-		{
-			BatchId = "cancel-member-batch",
-			State = JobState.AwaitingContinuation,
-			RemainingDependencies = 1,
-		};
-		var completeChild = CreateJob("cancel-complete-child", now) with
-		{
-			BatchId = "cancel-member-batch",
-			State = JobState.AwaitingContinuation,
-			RemainingDependencies = 1,
-		};
-		await storage.EnqueueBatchAsync(new()
-		{
-			Id = "cancel-member-batch",
-			CreatedAt = now,
-			TotalJobs = 3,
-			PendingCount = 3,
-			State = BatchState.Executing,
-		}, [parent, successChild, completeChild],
-		[
-			new()
-			{
-				ChildJobId = successChild.Id,
-				ParentJobId = parent.Id,
-				Trigger = ContinuationTrigger.Success,
-			},
-			new()
-			{
-				ChildJobId = completeChild.Id,
-				ParentJobId = parent.Id,
-				Trigger = ContinuationTrigger.Complete,
-			},
-		], cancellationToken);
-
-		await storage.CancelAsync(parent.Id, cancellationToken);
-
-		Assert.Equal(JobState.Cancelled, (await storage.GetJobStatusAsync(parent.Id, cancellationToken))!.State);
-		Assert.Equal(JobState.Skipped, (await storage.GetJobStatusAsync(successChild.Id, cancellationToken))!.State);
-		Assert.Equal(JobState.Pending, (await storage.GetJobStatusAsync(completeChild.Id, cancellationToken))!.State);
-		var status = Assert.IsType<BatchStatus>(await storage.GetBatchStatusAsync("cancel-member-batch", cancellationToken));
-		Assert.Equal(BatchState.Executing, status.State);
-		Assert.Equal(1, status.Cancelled);
-		Assert.Equal(1, status.Skipped);
-		Assert.Equal(1, status.Remaining);
-	}
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task SkippedFailureBranchDoesNotCancelSuccessfulBatch(
-		DatabaseKind database,
-		AdapterKind adapter
-	)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		var storage = fixture.GraphStorage;
-		var now = fixture.TimeProvider.GetUtcNow();
-		var parent = CreateJob("successful-parent", now) with { BatchId = "skipped-branch-batch" };
-		var failureOnly = CreateJob("failure-only", now) with
-		{
-			BatchId = "skipped-branch-batch",
-			State = JobState.AwaitingContinuation,
-			RemainingDependencies = 1,
-		};
-		await storage.EnqueueBatchAsync(new()
-		{
-			Id = "skipped-branch-batch",
-			CreatedAt = now,
-			TotalJobs = 2,
-			PendingCount = 2,
-			State = BatchState.Executing,
-		}, [parent, failureOnly], [new()
-		{
-			ChildJobId = failureOnly.Id,
-			ParentJobId = parent.Id,
-			Trigger = ContinuationTrigger.Failure,
-		}], cancellationToken);
-		_ = Assert.Single(await storage.AcquireDueJobsAsync(CreateRequest("worker", 1), cancellationToken));
-
-		await storage.CompleteAsync(parent.Id, 1, "worker", cancellationToken);
-
-		Assert.Equal(JobState.Skipped, (await storage.GetJobStatusAsync(failureOnly.Id, cancellationToken))!.State);
-		var status = Assert.IsType<BatchStatus>(await storage.GetBatchStatusAsync("skipped-branch-batch", cancellationToken));
-		Assert.Equal(BatchState.Succeeded, status.State);
-		Assert.Equal(1, status.Succeeded);
-		Assert.Equal(1, status.Skipped);
-		Assert.Equal(0, status.Cancelled);
-		Assert.Equal(0, status.Remaining);
-	}
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task EmptyBatchProgressIsFullySettled(DatabaseKind database, AdapterKind adapter)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		await fixture.CreateEmptyBatchAsync("empty-batch", cancellationToken);
-
-		var status = Assert.IsType<BatchStatus>(await fixture.GraphStorage.GetBatchStatusAsync(
-			"empty-batch",
-			cancellationToken
-		));
-
-		Assert.Equal(0, status.Total);
-		Assert.Equal(0, status.Remaining);
-		Assert.Equal(1d, status.FractionSettled);
-	}
-
-	[Theory]
-	[MemberData(nameof(Matrix))]
-	public async Task MissingBatchDashboardActionsThrowKeyNotFoundException(
-		DatabaseKind database,
-		AdapterKind adapter
-	)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, adapter, cancellationToken);
-		var graphStorage = fixture.GraphStorage;
-
-		_ = await Assert.ThrowsAsync<KeyNotFoundException>(
-			() => graphStorage.CancelBatchAsync("missing", cancellationToken).AsTask()
-		);
-		_ = await Assert.ThrowsAsync<KeyNotFoundException>(
-			() => graphStorage.DeleteBatchAsync("missing", cancellationToken).AsTask()
-		);
-	}
-
-	[Theory]
-	[InlineData(DatabaseKind.Sqlite)]
-	[InlineData(DatabaseKind.PostgreSql)]
-	[InlineData(DatabaseKind.SqlServer)]
-	public async Task AdaptersShareLinqToDBCreatedSchema(DatabaseKind database)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, AdapterKind.LinqToDB, cancellationToken);
-		var efStorage = fixture.CreateEntityFrameworkCoreStorage();
-		var now = fixture.TimeProvider.GetUtcNow();
-		await fixture.Storage.EnqueueAsync(CreateJob("from-linq2db", now), cancellationToken);
-		Assert.Equal("from-linq2db", Assert.Single(await efStorage.QueryJobsAsync(
-			new() { Id = "from-linq2db" },
-			cancellationToken
-		)).Id);
-		await efStorage.EnqueueAsync(CreateJob("from-ef", now.AddTicks(1)), cancellationToken);
-		Assert.Equal("from-ef", Assert.Single(await fixture.Storage.QueryJobsAsync(
-			new() { Id = "from-ef" },
-			cancellationToken
-		)).Id);
-	}
-
-	[Theory]
-	[InlineData(DatabaseKind.Sqlite)]
-	[InlineData(DatabaseKind.PostgreSql)]
-	[InlineData(DatabaseKind.SqlServer)]
-	public async Task AdaptersShareEntityFrameworkCoreCreatedSchema(DatabaseKind database)
-	{
-		var cancellationToken = TestContext.Current.CancellationToken;
-		await using var fixture = await CreateFixtureAsync(database, AdapterKind.EntityFrameworkCore, cancellationToken);
-		var linqStorage = fixture.CreateLinqToDBStorage();
-		var now = fixture.TimeProvider.GetUtcNow();
-		await fixture.Storage.EnqueueAsync(CreateJob("from-ef", now), cancellationToken);
-		Assert.Equal("from-ef", Assert.Single(await linqStorage.QueryJobsAsync(
-			new() { Id = "from-ef" },
-			cancellationToken
-		)).Id);
-		await linqStorage.EnqueueAsync(CreateJob("from-linq2db", now.AddTicks(1)), cancellationToken);
-		Assert.Equal("from-linq2db", Assert.Single(await fixture.Storage.QueryJobsAsync(
-			new() { Id = "from-linq2db" },
-			cancellationToken
-		)).Id);
-	}
-
-	private async Task<MatrixFixture> CreateFixtureAsync(
-		DatabaseKind database,
-		AdapterKind adapter,
-		CancellationToken cancellationToken
-	)
-	{
-		var schema = database == DatabaseKind.Sqlite ? null : "jobs_" + Guid.NewGuid().ToString("N");
-		var sqlitePath = database == DatabaseKind.Sqlite
-			? Path.Combine(Path.GetTempPath(), $"immediate-jobs-matrix-{Guid.NewGuid():N}.db")
-			: null;
-		var connectionString = database switch
-		{
-			DatabaseKind.Sqlite => $"Data Source={sqlitePath}",
-			DatabaseKind.PostgreSql => containers.PostgreSql.GetConnectionString(),
-			DatabaseKind.SqlServer => containers.SqlServer.GetConnectionString(),
-			_ => throw new ArgumentOutOfRangeException(nameof(database)),
-		};
-		var contextOptions = new DbContextOptionsBuilder<MatrixDbContext>();
-		DataOptions dataOptions;
-		if (database == DatabaseKind.Sqlite)
-		{
-			dataOptions = new DataOptions().UseSQLite(connectionString);
-			_ = contextOptions.UseSqlite(connectionString);
-		}
-		else if (database == DatabaseKind.PostgreSql)
-		{
-			dataOptions = new DataOptions().UsePostgreSQL(connectionString);
-			_ = contextOptions.UseNpgsql(connectionString);
-		}
-		else
-		{
-			dataOptions = new DataOptions().UseSqlServer(connectionString);
-			_ = contextOptions.UseSqlServer(connectionString);
-		}
-
-		_ = contextOptions.ReplaceService<IModelCacheKeyFactory, SchemaModelCacheKeyFactory>();
-		var contextFactory = new MatrixDbContextFactory(contextOptions.Options, schema);
-		var fixture = new MatrixFixture(
-			database,
-			connectionString,
-			schema,
-			sqlitePath,
-			dataOptions,
-			contextFactory,
-			adapter
-		);
-		try
-		{
-			if (adapter == AdapterKind.LinqToDB)
-			{
-				await dataOptions.CreateImmediateJobsSchemaAsync(schema, cancellationToken);
-			}
-			else
-			{
-				await using var context = contextFactory.CreateDbContext();
-				var script = context.Database.GenerateCreateScript();
-				foreach (var batch in Regex.Split(script, @"^\s*GO\s*$", RegexOptions.Multiline | RegexOptions.IgnoreCase, TimeSpan.FromSeconds(1)))
-				{
-					if (!string.IsNullOrWhiteSpace(batch))
-						_ = await context.Database.ExecuteSqlRawAsync(batch, cancellationToken);
-				}
-			}
-
-			return fixture;
-		}
-		catch
-		{
-			await fixture.DisposeAsync();
-			throw;
-		}
-	}
-
-	private static JobAcquisitionRequest CreateRequest(string workerId, int batchSize) => new()
-	{
-		WorkerId = workerId,
-		Lease = TimeSpan.FromMinutes(1),
-		BatchSize = batchSize,
-		Queues = [new()
-		{
-			QueueName = JobQueueDefinition.DefaultName,
-			Capacity = batchSize,
-			JobCapacities = new Dictionary<string, int> { ["matrix-test"] = batchSize },
-		}],
-	};
-
-	private static JobRecord CreateJob(string id, DateTimeOffset now) => new()
-	{
-		Id = id,
-		JobName = "matrix-test",
-		Payload = "{}",
-		State = JobState.Pending,
-		DueAt = now,
-		CreatedAt = now,
-	};
-
-	public enum DatabaseKind
-	{
-		Sqlite,
-		PostgreSql,
-		SqlServer,
-	}
-
-	public enum AdapterKind
-	{
-		EntityFrameworkCore,
-		LinqToDB,
-	}
-
-	private sealed class MatrixFixture(
-		DatabaseKind database,
-		string connectionString,
-		string? schema,
-		string? sqlitePath,
-		DataOptions dataOptions,
-		MatrixDbContextFactory contextFactory,
-		AdapterKind adapter
-	) : IAsyncDisposable
-	{
-		private readonly List<IJobStorage> _storages = [];
-
-		public FakeTimeProvider TimeProvider { get; } = new(new DateTimeOffset(2026, 7, 22, 10, 0, 0, TimeSpan.Zero));
-		public IJobStorage Storage => adapter == AdapterKind.LinqToDB
-			? CreateLinqToDBStorage()
-			: CreateEntityFrameworkCoreStorage();
-		public IRecurringJobStorage RecurringStorage => (IRecurringJobStorage)Storage;
-		public IJobGraphStorage GraphStorage => (IJobGraphStorage)Storage;
-
-		public IJobStorage CreateStorage() => Storage;
-
-		public async ValueTask CreateEmptyBatchAsync(string batchId, CancellationToken cancellationToken)
-		{
-			var storage = GraphStorage;
-			var now = TimeProvider.GetUtcNow();
-			var member = CreateJob($"{batchId}-member", now) with
-			{
-				BatchId = batchId,
-				State = JobState.Succeeded,
-				CompletedAt = now,
-			};
-			await storage.EnqueueBatchAsync(new()
-			{
-				Id = batchId,
-				CreatedAt = now,
-				TotalJobs = 1,
-				PendingCount = 0,
-				SucceededCount = 1,
-				State = BatchState.Succeeded,
-			}, [member], [], cancellationToken);
-
-			var tablePrefix = database switch
-			{
-				DatabaseKind.Sqlite => string.Empty,
-				DatabaseKind.PostgreSql => $"\"{schema}\".",
-				DatabaseKind.SqlServer => $"[{schema}].",
-				_ => throw new InvalidOperationException($"Unknown matrix database '{database}'."),
-			};
-			var batchTable = tablePrefix + (database == DatabaseKind.SqlServer
-				? "[immediate_job_batches]"
-				: "\"immediate_job_batches\"");
-			var jobTable = tablePrefix + (database == DatabaseKind.SqlServer
-				? "[immediate_jobs]"
-				: "\"immediate_jobs\"");
-			string Quote(string identifier) => database == DatabaseKind.SqlServer
-				? $"[{identifier}]"
-				: $"\"{identifier}\"";
-			await using var connection = new DataConnection(dataOptions);
-			_ = await connection.ExecuteAsync(
-				$"DELETE FROM {jobTable} WHERE {Quote("BatchId")} = '{batchId}'",
-				cancellationToken
-			);
-			_ = await connection.ExecuteAsync(
-				$"""
-				UPDATE {batchTable}
-				SET {Quote("TotalJobs")} = 0,
-					{Quote("PendingCount")} = 0,
-					{Quote("SucceededCount")} = 0,
-					{Quote("FailedCount")} = 0,
-					{Quote("CancelledCount")} = 0,
-					{Quote("SkippedCount")} = 0
-				WHERE {Quote("Id")} = '{batchId}'
-				""",
-				cancellationToken
-			);
-		}
-
-		public async ValueTask<int> CountServersAsync(CancellationToken cancellationToken)
-		{
-			var tablePrefix = database switch
-			{
-				DatabaseKind.Sqlite => string.Empty,
-				DatabaseKind.PostgreSql => $"\"{schema}\".",
-				DatabaseKind.SqlServer => $"[{schema}].",
-				_ => throw new InvalidOperationException($"Unknown matrix database '{database}'."),
-			};
-			var serverTable = tablePrefix + (database == DatabaseKind.SqlServer
-				? "[immediate_job_servers]"
-				: "\"immediate_job_servers\"");
-			await using var connection = new DataConnection(dataOptions);
-			return await connection.ExecuteAsync<int>(
-				$"SELECT COUNT(*) FROM {serverTable}",
-				cancellationToken
-			);
-		}
-
-		public EntityFrameworkCoreJobStorage<MatrixDbContext> CreateEntityFrameworkCoreStorage()
-		{
-			var storage = new EntityFrameworkCoreJobStorage<MatrixDbContext>(contextFactory, TimeProvider);
-			_storages.Add(storage);
-			return storage;
-		}
-
-		public LinqToDBJobStorage CreateLinqToDBStorage()
-		{
-			var storage = new LinqToDBJobStorage(dataOptions, schema, TimeProvider);
-			_storages.Add(storage);
-			return storage;
-		}
-
-		public async ValueTask DisposeAsync()
-		{
-			foreach (var storage in _storages.AsEnumerable().Reverse())
-				await storage.DisposeAsync();
-
-			if (sqlitePath is not null)
-			{
-				SqliteConnection.ClearAllPools();
-				File.Delete(sqlitePath);
-				return;
-			}
-
-			var cleanupOptions = database == DatabaseKind.PostgreSql
-				? new DataOptions().UsePostgreSQL(connectionString)
-				: new DataOptions().UseSqlServer(connectionString);
-			await using var connection = new DataConnection(cleanupOptions);
-			if (database == DatabaseKind.PostgreSql)
-			{
-				_ = await connection.ExecuteAsync($"DROP SCHEMA IF EXISTS \"{schema}\" CASCADE");
-			}
-			else
-			{
-				foreach (var table in new[]
-				{
-					"immediate_job_continuations",
-					"immediate_job_executions",
-					"immediate_fair_queue_groups",
-					"immediate_jobs",
-					"immediate_job_batches",
-					"immediate_recurring_jobs",
-					"immediate_job_servers",
-				})
-				{
-					_ = await connection.ExecuteAsync($"DROP TABLE IF EXISTS [{schema}].[{table}]");
-				}
-
-				_ = await connection.ExecuteAsync(
-					$"IF SCHEMA_ID(N'{schema}') IS NOT NULL EXEC(N'DROP SCHEMA [{schema}]')"
-				);
-			}
-		}
-
-	}
-
-	private sealed class MatrixDbContextFactory(
-		DbContextOptions<MatrixDbContext> options,
-		string? schema
-	) : IDbContextFactory<MatrixDbContext>
-	{
-		public MatrixDbContext CreateDbContext() => new(options, schema);
-	}
-
-	private sealed class MatrixDbContext(DbContextOptions<MatrixDbContext> options, string? schema)
-		: DbContext(options)
-	{
-		public string? Schema { get; } = schema;
-
-		protected override void OnModelCreating(ModelBuilder modelBuilder)
-		{
-			_ = modelBuilder.AddImmediateJobs(Schema);
-		}
-	}
-
-	private sealed class SchemaModelCacheKeyFactory : IModelCacheKeyFactory
-	{
-		public object Create(DbContext context, bool designTime) =>
-			(context.GetType(), ((MatrixDbContext)context).Schema, designTime);
-	}
-}
