+++ tests/Immediate.Jobs.StorageTests/StorageConformanceTests.cs
@@ -25,17 +25,23 @@ public sealed class RelationalStorageConformanceTests(StorageContainers containe
 		StorageCapabilities.FairQueues |
 		StorageCapabilities.Replica;
 
-	public static TheoryData<ConformanceDatabase, ConformanceAdapter, JobStorageConformanceTestCase> Cases
+	public static TheoryData<ConformanceDatabase, ConformanceAdapter, ConformanceTopology, JobStorageConformanceTestCase> Cases
 	{
 		get
 		{
-			var data = new TheoryData<ConformanceDatabase, ConformanceAdapter, JobStorageConformanceTestCase>();
+			var data = new TheoryData<ConformanceDatabase, ConformanceAdapter, ConformanceTopology, JobStorageConformanceTestCase>();
 			foreach (var database in Enum.GetValues<ConformanceDatabase>())
 			{
 				foreach (var adapter in Enum.GetValues<ConformanceAdapter>())
 				{
-					foreach (var testCase in JobStorageConformanceSuite.GetCases(Capabilities))
-						data.Add(database, adapter, testCase);
+					foreach (var topology in Enum.GetValues<ConformanceTopology>())
+					{
+						var capabilities = topology == ConformanceTopology.Distributed
+							? Capabilities
+							: Capabilities & ~StorageCapabilities.Replica;
+						foreach (var testCase in JobStorageConformanceSuite.GetCases(capabilities))
+							data.Add(database, adapter, topology, testCase);
+					}
 				}
 			}
 
@@ -48,6 +54,7 @@ public sealed class RelationalStorageConformanceTests(StorageContainers containe
 	public async Task RelationalStorageConforms(
 		ConformanceDatabase database,
 		ConformanceAdapter adapter,
+		ConformanceTopology topology,
 		JobStorageConformanceTestCase testCase
 	)
 	{
@@ -56,30 +63,8 @@ public sealed class RelationalStorageConformanceTests(StorageContainers containe
 			containers,
 			database,
 			adapter,
-			TestContext.Current.CancellationToken
-		);
-		await testCase.RunAsync(fixture.Services, TestContext.Current.CancellationToken);
-	}
-}
-
-public sealed class SingleServerStorageConformanceTests
-{
-	private const StorageCapabilities Capabilities =
-		StorageCapabilities.Queue |
-		StorageCapabilities.Recurring |
-		StorageCapabilities.Graph |
-		StorageCapabilities.FairQueues;
-
-	public static TheoryData<JobStorageConformanceTestCase> Cases =>
-		[.. JobStorageConformanceSuite.GetCases(Capabilities)];
-
-	[Theory]
-	[MemberData(nameof(Cases))]
-	public async Task SingleServerStorageConforms(JobStorageConformanceTestCase testCase)
-	{
-		ArgumentNullException.ThrowIfNull(testCase);
-		await using var fixture = await RelationalConformanceFixture.CreateSingleServerAsync(
-			TestContext.Current.CancellationToken
+			TestContext.Current.CancellationToken,
+			useDistributedTopology: topology == ConformanceTopology.Distributed
 		);
 		await testCase.RunAsync(fixture.Services, TestContext.Current.CancellationToken);
 	}
@@ -119,6 +104,12 @@ public enum ConformanceAdapter
 {
 	EntityFrameworkCore,
 	LinqToDB,
+}
+
+public enum ConformanceTopology
+{
+	Distributed,
+	SingleServer,
 }
 
 internal sealed class RelationalConformanceFixture : IAsyncDisposable
@@ -248,16 +239,6 @@ internal sealed class RelationalConformanceFixture : IAsyncDisposable
 			throw;
 		}
 	}
-
-	internal static ValueTask<RelationalConformanceFixture> CreateSingleServerAsync(
-		CancellationToken cancellationToken
-	) => CreateAsync(
-		containers: null,
-		ConformanceDatabase.Sqlite,
-		ConformanceAdapter.EntityFrameworkCore,
-		cancellationToken,
-		useDistributedTopology: false
-	);
 
 	private static StorageContainers GetContainers(StorageContainers? containers) =>
 		containers ?? throw new InvalidOperationException("A container fixture is required for server databases.");
